+++ apps/web/src/App.tsx
@@ -4,6 +4,7 @@ import { BillingPage } from './pages/BillingPage'
 import { EvidencePage } from './pages/EvidencePage'
 import { FindingsPage } from './pages/FindingsPage'
 import { HomePage } from './pages/HomePage'
+import { LoginPage } from './pages/LoginPage'
 import { OperationsPage } from './pages/OperationsPage'
 import { OverviewPage } from './pages/OverviewPage'
 import { ReportsPage } from './pages/ReportsPage'
@@ -11,6 +12,7 @@ import { ReportsPage } from './pages/ReportsPage'
 export function App() {
   return (
     <Routes>
+      <Route path="login" element={<LoginPage />} />
       <Route element={<AppShell />}>
         <Route index element={<HomePage />} />
         <Route path="overview" element={<OverviewPage />} />